+++ src/Language/Haskell/Liquid/CTags.hs
@@ -28,7 +28,7 @@ import qualified Data.Graph             as G
 
 import Language.Fixpoint.Misc         (mapSnd)
 import Language.Fixpoint.Types     (Tag)
-import Language.Haskell.Liquid.GhcInterface (freeVars)
+import Language.Haskell.Liquid.Visitors (freeVars)
 
 -- | The @TagKey@ is the top-level binder, and @Tag@ is a singleton Int list
