+++ src/Language/Haskell/Liquid/DiffCheck.hs
@@ -42,8 +42,9 @@ import            System.Directory                (copyFile, doesFileExist)
 import            Language.Fixpoint.Types         (FixResult (..))
 import            Language.Fixpoint.Files
 import            Language.Haskell.Liquid.Types   (AnnInfo (..), Error, TError (..), Output (..))
-import            Language.Haskell.Liquid.GhcInterface
 import            Language.Haskell.Liquid.GhcMisc
+import            Language.Haskell.Liquid.Visitors
+import            Language.Haskell.Liquid.Errors   ()
 import            Text.Parsec.Pos                  (sourceName, sourceLine, sourceColumn, SourcePos, newPos)
 import            Text.PrettyPrint.HughesPJ        (text, render, Doc)
