+++ tests/pos/Strings.hs
@@ -1,0 +1,8 @@
+module DBC where
+
+import GHC.CString  -- This import interprets Strings as constants!
+
+
+bar :: () -> String 
+{-@ bar :: () -> {x:String | x = foo} @-}
+bar _ = "foo"
